Fixed some bugs and improved algorithm. Added visual components.
Cleaned up majority of bugs, implemented the algorithm feature. Lots of bugs fixed, lots more still to be worked out. Known bugs:
if food spawns on same level (X or Y) as the snake it will only generate paths directly to it. Right now the algorithm is still considering snakes a playable space.